third_party/WebKit/LayoutTests/fast/notifications/notifications-leak-document-expected.txt

DESKTOP NOTIFICATION: icon , title New E-mail, text Meet me tonight at 8!
DESKTOP NOTIFICATION CLOSED: New E-mail
PASS The difference of counter "numberOfLiveDocuments" before and after the cycle is under the threshold of 0.
PASS successfullyParsed is true

TEST COMPLETE
